+++ src/Light.cpp
@@ -21,8 +21,8 @@ Light::Light(   GLfloat red, GLfloat green, GLfloat blue, GLfloat ambIntensity,
     diffuseIntensity = diffIntensity;
 }
 
-void Light::useLight(   GLfloat ambientIntensityLocation, GLfloat ambientColourLocation,
-                        GLfloat diffuseIntensityLocation, GLfloat directionLocation  ) {
+void Light::useLight(   GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+                        GLuint diffuseIntensityLocation, GLuint directionLocation  ) {
     // Ambient Light
     glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
